move vops to own file, add vec/ and note the perf changes vs vec+
Awesome stuff here, before the summing of lisp arrays was pretty
competitive (only 2x as slow as sse). I had seen something like this
in a gdc talk on SSE[0]. The reason was that whilst the `add` µop had
a latency of 3 cycles, it had a throughput of 1 cycle, so every cycle
a new `add` could be dispatched. SSE's `add` has (IIRC) a latency and
throughput of 3 cycles so you have to wait for one to finish before
dispatching the next.

To counter this I chose divide instead as, in the video, divide had a
throughput that matched the latency. With this change the SSE code
performed a near perfect 4x faster than the lisp code.

Hopefully this wasnt just some other factor I dont understand yet as
I'm kind of stoked to see this nice behaviour in the code.

That's all the malinformed play for tonight. Seeya!

> [0] http://gdcvault.com/play/1014645/Hotspots-FLOPS-and-uOps-To